+++ sqlite3/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2
+
+- Enable extended result codes
+- Expose raw rows from a `ResultSet`
+
 ## 0.1.1
 
 - Expose the `ResultSet` class